Fix bug in distribution.

+++ scripts/simulations.py
@@ -44,14 +44,16 @@ def generate_data_and_true_distr(path, seeds, samp_size_range, num_levels_range)
         cards = [2] * num_levels
         for samp_size in samp_size_range:
             for seed in seeds:
+                
+                if Path(f"{data_path}/p={num_levels}_n={samp_size}_seed={seed}.csv").is_file():
+                    continue
                 print(f"Starting run {seed} for p={num_levels}...")
-
                 np.random.seed(seed)
                 random.seed(seed)
 
                 # generate random cstree and compute its kl divergence
                 tree = ct.sample_cstree(cards, max_cvars=2, prob_cvar=0.5, prop_nonsingleton=1)
-                tree.sample_stage_parameters(alpha=2)
+                tree.sample_stage_parameters(alpha=1)
                 data = tree.sample(samp_size)
                 tree_df = tree.to_joint_distribution(label_order=list(data.columns))
                 tree_df.to_csv(f"{distr_path}/p={num_levels}_n={samp_size}_seed={seed}.csv", index=False)
@@ -69,10 +71,12 @@ def estimate_cstree_distr(data_path, est_path, seeds, samp_size_range, num_level
         for samp_size in samp_size_range:
             for seed in seeds:
                 name = f"p={num_levels}_n={samp_size}_seed={seed}.csv"
-    
-                print(f"Estimating CStree for {data_path}/{name}...")
+                np.random.seed(seed)
+                random.seed(seed)
+                
                 if Path(f"{est_path}/{name}").is_file():
                     continue
+                print(f"Estimating CStree for {data_path}/{name}...")
                 data = pd.read_csv(f"{data_path}/{name}")
                 
                 score_table, context_scores, context_counts = sc.order_score_tables(
@@ -80,7 +84,7 @@ def estimate_cstree_distr(data_path, est_path, seeds, samp_size_range, num_level
                 orders, scores = ctl.gibbs_order_sampler(5000, score_table)
                 maporder = orders[scores.index(max(scores))]
                 tree = ctl._optimal_cstree_given_order(maporder, context_scores)
-                tree.estimate_stage_parameters(data)
+                tree.estimate_stage_parameters(data, alpha_tot=1.0)
 
                 tree_df = tree.to_joint_distribution(label_order=list(data.columns))
                 tree_df.to_csv(f"{est_path}/{name}", index=False)
@@ -90,17 +94,23 @@ def kl_div_from_files(true_path, est_path, seeds, samp_size_range, num_levels_ra
     """ Read in true and estimated distributions from files and compute KL divergence.
     """
     # get all distribution files
-    kl_divs = []
+    
     for num_levels in num_levels_range:
         cards = [2] * num_levels
         for samp_size in samp_size_range:
+            kl_divs = []
             for seed in seeds:
+                
                 name = f"p={num_levels}_n={samp_size}_seed={seed}.csv"
+                #print(f"Computing KL divergence for {name}...")
                 true_dist_df = pd.read_csv(f"{true_path}/{name}")
                 est_dist_df = pd.read_csv(f"{est_path}/{name}")
                 
                 kl_divs.append(KL_divergence(true_dist_df, est_dist_df))
-    return np.array(kl_divs)
+            name = f"p={num_levels}_n={samp_size}"
+            print(f"KL divergence for {name}:")
+            #print(kl_divs)
+            print(f"mean:{np.array(kl_divs).mean():.2f} std:{np.array(kl_divs).std():.2f}")
 
 if __name__ == "__main__":
     # check versions to ensure accurate reproduction
@@ -110,19 +120,18 @@ if __name__ == "__main__":
     path = "sim_results"
     warnings.simplefilter(action="ignore", category=FutureWarning)
 
-    samp_size_range = [1000]
-    seeds = list(range(10))
-    num_levels_range = [5, 7]
+    samp_size_range = [100, 1000, 10000]
+    seeds = list(range(20))
+    num_levels_range = [5]
 
     generate_data_and_true_distr(path, seeds, samp_size_range, num_levels_range)
     estimate_cstree_distr(f"{path}/data", f"{path}/distr/cstree_est", seeds, samp_size_range, num_levels_range)
-    #kldivs = kl_div_from_files(f"{path}/distr/true", f"{path}/distr/cstree_est", seeds, samp_size_range, num_levels_range)
-    #kldivs2 = kl_div_from_files(f"{path}/distr/true", f"{path}/distr/stagedtrees", seeds, samp_size_range, num_levels_range)
-    kldivs = kl_div_from_files(f"{path}/distr/cstree_est", f"{path}/distr/true", seeds, samp_size_range, num_levels_range)
-    kldivs2 = kl_div_from_files(f"{path}/distr/stagedtrees", f"{path}/distr/true", seeds, samp_size_range, num_levels_range)
+    print("Estimated cstree KL")
+    kl_div_from_files(f"{path}/distr/cstree_est", f"{path}/distr/true", seeds, samp_size_range, num_levels_range)
+    print("Staged trees KL")
+    kl_div_from_files(f"{path}/distr/stagedtrees", f"{path}/distr/true", seeds, samp_size_range, num_levels_range)
     
-    
-    print(kldivs.mean(), kldivs.std())
-    print(kldivs2.mean(), kldivs2.std())
+    #print(kldivs.mean(), kldivs.std())
+    #print(kldivs2.mean(), kldivs2.std())
     os.environ.clear()
     os.environ.update(old_environ)
